pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
include = ["daiana*"]
exclude = ["tests*", "job_hunt*", "docs*"]
namespaces = false


[project]
name = "daiana"
version = "1.0.0"
description = "AI empowered tool for LaTeX CV/cover letter automation for job hunting"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "PanoPepino", email = "panizofisico@gmail.com" }
]
keywords = ["cli", "latex", "cv", "cover-letter", "job-hunting", "api-services", "AI"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]


dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pypdf>=3.0.0",
]


[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "responses>=0.25",
]
docs = [
    "sphinx>=7.0",
    "furo",
    "sphinx-design",
]


[project.scripts]
daiana = "daiana.cli:app"


[project.urls]
Homepage = "https://github.com/PanoPepino/Daiana"
Repository = "https://github.com/PanoPepino/Daiana"